// src/frontend/src/backend.d.ts
export type Time = bigint;
export interface Feedback {
    id: bigint;
    status: FeedbackStatus;
    title: string;
    createdAt: Time;
    authorName?: string;
    message: string;
    category: FeedbackCategory;
    authorPrincipal?: Principal;
}
// ...
export enum FeedbackCategory {
    bug = "bug",
    idea = "idea",
    issue = "issue"
}
export enum FeedbackStatus {
    pending = "pending",
    completed = "completed"
}